modules/fbs-sql-generator/pyproject.toml
[tool.poetry]
name = "datagen"
version = "0.1.0"
description = ""
authors = ["Jann Hendrik Wißbach <jann.hendrik.wissbach@mni.thm.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
faker = "^19.3.0"
mysql-connector-python = "^8.1.0"
datetime = "^5.5"
flask = {extras = ["async"], version = "^3.0.3"}
gunicorn = "^22.0.0"
dominate = "^2.9.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
